+++ packages/runtime/src/helpers/edge.ts
@@ -3,9 +3,14 @@ import { promises as fs, existsSync } from 'fs'
 import { resolve, join } from 'path'
 
 import type { NetlifyConfig, NetlifyPluginConstants } from '@netlify/build'
+import { greenBright } from 'chalk'
+import destr from 'destr'
 import { copy, copyFile, emptyDir, ensureDir, readJSON, readJson, writeJSON, writeJson } from 'fs-extra'
 import type { MiddlewareManifest } from 'next/dist/build/webpack/plugins/middleware-plugin'
 import type { RouteHas } from 'next/dist/lib/load-custom-routes'
+import { outdent } from 'outdent'
+
+import { getRequiredServerFiles } from './config'
 
 // This is the format as of next@12.2
 interface EdgeFunctionDefinitionV1 {
@@ -40,10 +45,12 @@ export interface FunctionManifest {
   functions: Array<
     | {
         function: string
+        name?: string
         path: string
       }
     | {
         function: string
+        name?: string
         pattern: string
       }
   >
@@ -222,9 +229,17 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
   const edgeFunctionRoot = resolve('.netlify', 'edge-functions')
   await emptyDir(edgeFunctionRoot)
 
+  const { publish } = netlifyConfig.build
+  const nextConfigFile = await getRequiredServerFiles(publish)
+  const nextConfig = nextConfigFile.config
   await copy(getEdgeTemplatePath('../edge-shared'), join(edgeFunctionRoot, 'edge-shared'))
+  await writeJSON(join(edgeFunctionRoot, 'edge-shared', 'nextConfig.json'), nextConfig)
 
-  if (!process.env.NEXT_DISABLE_EDGE_IMAGES) {
+  if (
+    !destr(process.env.NEXT_DISABLE_EDGE_IMAGES) &&
+    !destr(process.env.NEXT_DISABLE_NETLIFY_EDGE) &&
+    !destr(process.env.DISABLE_IPX)
+  ) {
     console.log(
       'Using Netlify Edge Functions for image format detection. Set env var "NEXT_DISABLE_EDGE_IMAGES=true" to disable.',
     )
@@ -240,15 +255,17 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
       path: '/_next/image*',
     })
   }
-
-  if (!process.env.NEXT_DISABLE_NETLIFY_EDGE) {
+  if (!destr(process.env.NEXT_DISABLE_NETLIFY_EDGE)) {
     const middlewareManifest = await loadMiddlewareManifest(netlifyConfig)
     if (!middlewareManifest) {
       console.error("Couldn't find the middleware manifest")
       return
     }
 
+    let usesEdge = false
+
     for (const middleware of middlewareManifest.sortedMiddleware) {
+      usesEdge = true
       const edgeFunctionDefinition = middlewareManifest.middleware[middleware]
       const functionDefinitions = await writeEdgeFunction({
         edgeFunctionDefinition,
@@ -261,6 +278,7 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
     // No, the version field was not incremented
     if (typeof middlewareManifest.functions === 'object') {
       for (const edgeFunctionDefinition of Object.values(middlewareManifest.functions)) {
+        usesEdge = true
         const functionDefinitions = await writeEdgeFunction({
           edgeFunctionDefinition,
           edgeFunctionRoot,
@@ -272,7 +290,7 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
 
     if (usesEdgeRouter()) {
       console.log('Using experimental Netlify Next.js Edge Router')
-      await writeEdgeRouter({ edgeFunctionRoot, publishDir: netlifyConfig.build.publish })
+      await writeEdgeRouter({ edgeFunctionRoot, publishDir: publish })
       // Pre-middleware routing runs first...
       manifest.functions.unshift({
         function: EDGE_ROUTER_PRE_MIDDLEWARE,
@@ -284,6 +302,12 @@ export const writeEdgeFunctions = async (netlifyConfig: NetlifyConfig) => {
         function: EDGE_ROUTER_POST_MIDDLEWARE,
         path: '/*',
       })
+    }
+    if (usesEdge) {
+      console.log(outdent`
+        ✨ Deploying middleware and functions to ${greenBright`Netlify Edge Functions`} ✨
+        This feature is in beta. Please share your feedback here: https://ntl.fyi/next-netlify-edge
+      `)
     }
   }
   await writeJson(join(edgeFunctionRoot, 'manifest.json'), manifest)